update profile pages
removed buttons roe editing and replaced with a button and a bottom sheet.
added variable color to discription text field and amount tiem check box field.
added check box helper to make the check box function within the bottom sheet.

the dropdown widgets need to be compleated with services.
the jobs widgets needs to be compleated with services